move common utils to sajari-react-sdk-utils package
