+++ packages/utils/tsdx.config.js
@@ -1,0 +1,5 @@
+const base = require('../../tsdx.config.base.js');
+
+module.exports = {
+  ...base,
+};
